Add HerokuSSL rack middleware
Without this patch the puppet master running on heroku returns the
Puppet CA self-signed root CA certificate when the agent requests the CA
certificate.  This is a problem because this CA certificate has not
issued the SSL certificate for *.herokuapp.com

This patch addresses the problem by using a Rack middleware to intercept
the REST API request for the CA certificate and returning one
appropriate for use with Heroku.

NOTE: There can only be 1 certificate in the response, thankfully
Heroku's load balancer fills in the rest of the certificate chain for us
so we need only add the root CA certificate for DigiSign in order to get
the agent working correctly.